# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API backend",
        "description": "Documentação da API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projetos": {
            "get": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Listar todos os projetos",
                "operationId": "88e1f5b58f31f78ac4f9c29c51b2b0ba",
                "responses": {
                    "200": {
                        "description": "Lista de projetos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Projeto"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Criar um novo projeto",
                "operationId": "b98fd48a5abec33d26e9af7abd8e0b02",
                "requestBody": {
                    "description": "Dados para criar um novo projeto, incluindo o arquivo binário",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "arquivo",
                                    "nome_arquivo",
                                    "descricao",
                                    "id_empresa"
                                ],
                                "properties": {
                                    "arquivo": {
                                        "description": "O arquivo binário do projeto (ex: PDF, imagem). Use form-data.",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nome_arquivo": {
                                        "description": "Nome original do arquivo enviado",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição detalhada do projeto",
                                        "type": "string"
                                    },
                                    "versao": {
                                        "description": "Versão do projeto (opcional)",
                                        "type": "string"
                                    },
                                    "ativo": {
                                        "description": "Status de ativação do projeto (opcional, padrão true)",
                                        "type": "boolean"
                                    },
                                    "id_empresa": {
                                        "description": "ID da empresa à qual o projeto pertence",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projeto criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projeto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/projetos/{id}": {
            "get": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Exibir um projeto específico",
                "operationId": "a7bb6e38f95809182a60fea7624e2c87",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID do Projeto",
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projeto encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projeto"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Atualizar um projeto existente",
                "operationId": "84b63cdbc13b927614038111d71740cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID do Projeto",
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados para atualizar um projeto (todos os campos são opcionais)",
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "arquivo": {
                                        "description": "Um novo arquivo binário do projeto (opcional, para substituição).",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "nome_arquivo": {
                                        "description": "Nome atualizado do arquivo",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição atualizada do projeto.",
                                        "type": "string"
                                    },
                                    "versao": {
                                        "description": "Nova versão do projeto",
                                        "type": "string"
                                    },
                                    "ativo": {
                                        "description": "Novo status de ativação",
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projeto atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Projeto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado a atualizar este projeto"
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Projetos"
                ],
                "summary": "Remover um projeto",
                "operationId": "70b8200445ebf2b0ed35ffd8348b66f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "description": "ID do Projeto",
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Projeto removido com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Não autorizado a remover este projeto"
                    },
                    "404": {
                        "description": "Projeto não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Projeto": {
                "title": "Projeto",
                "description": "Detalhes de um Projeto",
                "required": [
                    "id",
                    "arquivo",
                    "nome_arquivo",
                    "descricao",
                    "data_criacao",
                    "id_empresa"
                ],
                "properties": {
                    "id": {
                        "description": "ID único do projeto",
                        "type": "integer",
                        "format": "int64",
                        "readOnly": true,
                        "example": 1
                    },
                    "arquivo": {
                        "description": "Caminho ou URL do arquivo do projeto armazenado (Ex: storage/projetos/documento.pdf)",
                        "type": "string",
                        "example": "storage/projetos/documento_projeto_final.pdf"
                    },
                    "nome_arquivo": {
                        "description": "Nome original do arquivo enviado",
                        "type": "string",
                        "example": "Relatório Anual 2024.pdf"
                    },
                    "descricao": {
                        "description": "Descrição detalhada do projeto",
                        "type": "string",
                        "example": "Este projeto abrange a otimização de processos internos e a expansão de mercado para o próximo ano."
                    },
                    "data_criacao": {
                        "description": "Data de criação do projeto (formato ISO 8601: YYYY-MM-DD)",
                        "type": "string",
                        "format": "date",
                        "example": "2025-05-28"
                    },
                    "versao": {
                        "description": "Versão do projeto (opcional)",
                        "type": "string",
                        "example": "1.0.1"
                    },
                    "ativo": {
                        "description": "Status de ativação do projeto (true se ativo, false se inativo)",
                        "type": "boolean",
                        "example": true
                    },
                    "id_empresa": {
                        "description": "ID da empresa à qual o projeto pertence",
                        "type": "integer",
                        "format": "int64",
                        "example": 5
                    },
                    "created_at": {
                        "description": "Timestamp de criação do registro",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Timestamp da última atualização do registro",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": " * @OA\\SecurityScheme(\n * securityScheme=\"sanctum\",\n * type=\"http\",\n * scheme=\"bearer\",\n * bearerFormat=\"Sanctum\"\n * )\n ",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}